+++ src/app/services/wizard.service.ts
@@ -37,7 +37,7 @@ export class WizardService {
    * The project that is built with the wizard
    */
   public builtProject: ProjectAdd = {
-    callToAction: undefined,
+    callToActions: [],
     collaborators: [],
     name: '',
     fileId: 0,
@@ -193,7 +193,7 @@ export class WizardService {
                     shortDescription: project.shortDescription,
                     name: project.name,
                     fileId: undefined,
-                    callToAction: project.callToAction,
+                    callToActions: project.callToActions,
                     uri: projectUri,
                     description: project.description
                   });
@@ -295,7 +295,7 @@ export class WizardService {
     this.selectedSource = undefined;
     this.uploadFile = undefined;
     this.builtProject = {
-      callToAction: undefined,
+      callToActions: undefined,
       collaborators: [],
       name: '',
       fileId: 0,
@@ -382,7 +382,7 @@ export class WizardService {
     updatedSteps.find(step => step.wizardPageName === 'project-name').isComplete = WizardService.checkNotEmpty(project.name);
     updatedSteps.find(step => step.wizardPageName === 'project-icon').isComplete = !!project.projectIcon;
     updatedSteps.find(step => step.wizardPageName === 'project-collaborators').isComplete = project.collaborators.length > 0;
-    updatedSteps.find(step => step.wizardPageName === 'project-call-to-action').isComplete = !!project.callToAction;
+    updatedSteps.find(step => step.wizardPageName === 'project-call-to-action').isComplete = !!project.callToActions;
     updatedSteps.find(step => step.wizardPageName === 'project-link').isComplete = !!project.uri;
     updatedSteps.find(step => step.wizardPageName === 'project-description').isComplete = (
         WizardService.checkNotEmpty(project.description) && WizardService.checkNotEmpty(project.shortDescription)