+++ CHANGELOG.md
@@ -11,6 +11,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Multiple call to actions supported. [#454](https://github.com/DigitalExcellence/dex-frontend/issues/454)
+
 ### Deprecated
 
 ### Removed